Implement ExpressJS

+++ src/lib/user/application/UserDelete/UserDelete.ts
@@ -1,4 +1,5 @@
 import { UserId } from "../../domain/UserId";
+import { UserNotFoundError } from "../../domain/UserNotFoundError";
 import { UserRepository } from "../../domain/UserRepository";
 
 export class UserDelete {
@@ -6,6 +7,11 @@ export class UserDelete {
     constructor(private userRepository: UserRepository) {}
 
     async run(id: string): Promise<void> {
+        const userExists = await this.userRepository.getOneById(new UserId(id));
+
+        if (!userExists) throw new UserNotFoundError('User not found!');
+
+
         return this.userRepository.delete(new UserId(id));
     }
 }
